Misc release prep cleanup
