+++ modules/core/src/test/java/org/apache/fluo/recipes/core/export/it/RefInfo.java
@@ -13,24 +13,14 @@
  * the License.
  */
 
-package org.apache.fluo.recipes.core.map;
+package org.apache.fluo.recipes.core.export.it;
 
-import java.util.Iterator;
-import java.util.Optional;
+class RefInfo {
+  long seq;
+  boolean deleted;
 
-public class WordCountCombiner implements Combiner<String, Long> {
-  @Override
-  public Optional<Long> combine(String key, Iterator<Long> updates) {
-    long sum = 0;
-
-    while (updates.hasNext()) {
-      sum += updates.next();
-    }
-
-    if (sum == 0) {
-      return Optional.empty();
-    } else {
-      return Optional.of(sum);
-    }
+  public RefInfo(long seq, boolean deleted) {
+    this.seq = seq;
+    this.deleted = deleted;
   }
 }
