+++ modules/core/src/test/java/org/apache/fluo/recipes/core/export/it/RefUpdates.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package org.apache.fluo.recipes.core.export;
+package org.apache.fluo.recipes.core.export.it;
 
 import java.util.Set;
